「四則演算」：「getIntExtra」から「getFloatExtra」に変換。 「初期値：0」から「初期値：0.0」に変換。

+++ app/src/main/java/jp/techacademy/yuki/naka/calcapp/SecondActivity.kt
@@ -11,7 +11,7 @@ class SecondActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_second)
 
-        val value = intent.getIntExtra("VALUE",0)
+        val value = intent.getDoubleExtra("VALUE",0.0)
 
         textView.text ="${value}"
     }
